+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+- **Enhancement**: Prevent underlying page scroll when the search window is open
 - **Bugfix**: Unsplash links now comply to [their API guidelines](https://help.unsplash.com/en/articles/2511245-unsplash-api-guidelines). (Refs #1)
 
 ## 2020-02-06 - v0.2.0